# Struct-valued custom enumerator for pixel offsets