--- a/Pinta.Effects/Effects/PolarInversionEffect.cs
+++ b/Pinta.Effects/Effects/PolarInversionEffect.cs
@@ -60,7 +60,7 @@ public sealed class PolarInversionEffect : BaseEffect
 		Span<ColorBgra> dst_data = dst.GetPixelData ();
 		ReadOnlySpan<ColorBgra> src_data = src.GetReadOnlyPixelData ();
 		foreach (RectangleI rect in rois) {
-			foreach (var pixel in Utility.GeneratePixelOffsets (rect, src.GetSize ())) {
+			foreach (var pixel in Tiling.GeneratePixelOffsets (rect, src.GetSize ())) {
 				dst_data[pixel.memoryOffset] = Warp.GetPixelColor (
 					settings,
 					InverseTransform,