I think one shouldn't have to allocate on the heap in order to enumerate the pixel coordinates and offsets.

To be consumed by a `foreach` loop an object doesn't need to implement `IEnumerable<T>`, but only expose public methods with the corresponding names and signatures.

It is my understanding that this new approach would prevent heap allocations except when the object is explicitly used as an `IEnumerable<T>` (like in `VoronoiDiagramEffect` through an invocation of `Select`).