@cameronwhite 2024-12-10
Could you try running the effect benchmarks to see if there's a notable difference with this change? If so I'm okay with it, but otherwise the original implementation is much simpler