@cameronwhite 2024-12-23
Getting back to this, I ran the benchmarks myself and there is some improvement to the overall performance, so I'm okay with the change since it's a core utility used by many effects.
Could you just add some doc comments to the new classes to explain their purpose?

Before:
```
| Method      | Mean     | Error   | StdDev  | Allocated |
|------------ |---------:|--------:|--------:|----------:|
| BulgeEffect | 208.1 ms | 1.67 ms | 1.39 ms |   1.14 KB |
```

After:
```
| Method      | Mean     | Error   | StdDev  | Allocated |
|------------ |---------:|--------:|--------:|----------:|
| BulgeEffect | 196.3 ms | 2.38 ms | 2.23 ms |   1.04 KB |
```