@Lehonti 2024-12-23
Thank you for running the benchmarks, @cameronwhite! I tried to run them but ran into problems when NUnit refused to run the tests because they referenced an unoptimized library from Mono, and since these last few weeks have been busy I didn't have the time to figure out what the problem was.

I just added documentation comments to the new `struct`s